Use GitVersion to figure out package version
The build script now pokes all fsproj files with a <PackageVersion> set
to 0.0.0 and replaces the version with one provided by the GitVersion
tool. At the end of the build, only those project files which have a
version matching that which GitVersion provides will be reset to 0.0.0.

This allows manually setting a different package version locally, if so
desired, while still allowing the build script to poke the project files
locally if the version number is not important, without resulting in
changes that could accidentally be checked into version control.